refactor: 컴포넌트 이름 index.tsx에서 ImageSlider로 수정
