+++ src/components/ImageSlider/ImageSlider.tsx
