+++ src/components/index.ts
@@ -1,3 +1,3 @@
 export { default as Header } from './Header';
 export { default as ImageSelector } from './ImageSelector';
-export { default as ImageSlider } from './ImageSlider';
+export { default as ImageSlider } from './ImageSlider/ImageSlider';
